added basic welding complexity
At the moment, welding is an all or nothing job. In the future, I was thinking that you could have the cost of repairing or attaching a part vary from part to part based on its complexity. Repairing something from light green to dark green should be less fuel and time consuming than from dark red to dark green. Perhaps in the welding program, a check to see the damage on the selected part would bring up a fraction that could be multiplied by the part's complexity to calculate how much welding it would take to repair that part, and in turn how much time and fuel. In addition, attaching these parts would be more difficult depending on how complex. Attaching a piece of frame to a car wouldn't be as hard as another engine, or a minireactor. Feel free to look over my values and check to see if anything doesn't agree with you. I used 50 as the base value for frames and boards and such.